IOS-7018 Grooming folders
IOS-7018 Update Package.resolved
