+++ Tangem/Modules/Send/Fee/Interactor/SendFeeInteractor.swift
@@ -1,5 +1,5 @@
 //
-//  CommonSendFeeInteractor.swift
+//  SendFeeInteractor.swift
 //  Tangem
 //
 //  Created by Sergey Balashov on 18.06.2024.
@@ -9,6 +9,20 @@
 import Foundation
 import Combine
 import BlockchainSdk
+
+protocol SendFeeInteractor {
+    var selectedFee: SendFee? { get }
+
+    func update(selectedFee: SendFee)
+    func updateFees()
+
+    func feesPublisher() -> AnyPublisher<[SendFee], Never>
+    func selectedFeePublisher() -> AnyPublisher<SendFee?, Never>
+
+    func customFeeInputFieldModels() -> [SendCustomFeeInputFieldModel]
+
+    func setup(input: SendFeeInput, output: SendFeeOutput)
+}
 
 class CommonSendFeeInteractor {
     private weak var input: SendFeeInput?
@@ -59,7 +73,7 @@ class CommonSendFeeInteractor {
             .compactMap { interactor, fees in
                 fees.value.flatMap { interactor.initialFeeForUpdate(fees: $0) }
             }
-            // Only once
+        // Only once
             .first()
             .withWeakCaptureOf(self)
             .sink { interactor, fee in
@@ -222,3 +236,4 @@ private extension CommonSendFeeInteractor {
         output?.feeDidChanged(fee: fee)
     }
 }
+
