+++ TangemApp.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "a0738665bda10cbdb05da546422ce445e2fdf9ace530c7941d4cba18d71a3d25",
+  "originHash" : "b4054eb42ba68d5494e329b1b5ada27b86a35abfe2422c9f838f1be5b357085c",
   "pins" : [
     {
       "identity" : "anyasyncsequence",
@@ -170,6 +170,15 @@
       "state" : {
         "revision" : "b9906e7e5c56706f0fe1ad3545dbf02dcf195b34",
         "version" : "4.0.21-tangem6"
+      }
+    },
+    {
+      "identity" : "zipfoundation",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/weichsel/ZIPFoundation.git",
+      "state" : {
+        "revision" : "b979e8b52c7ae7f3f39fa0182e738e9e7257eb78",
+        "version" : "0.9.18"
       }
     }
   ],
